Move retryable dashboard to nextjs
