+++ .env.sample
@@ -1,1 +1,1 @@
-REACT_APP_INFURA_KEY=
+NEXT_PUBLIC_INFURA_KEY=
