+++ next.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+  async redirects() {
+    return [
+      {
+        source: '/tx',
+        destination: '/',
+        permanent: true,
+      },
+    ];
+  },
+};
