+++ package.json
@@ -4,35 +4,26 @@
   "license": "Apache-2.0",
   "private": true,
   "dependencies": {
-    "@arbitrum/sdk": "^3.0.0",
+    "@arbitrum/sdk": "^3.1.1",
     "@ethersproject/address": "^5.0.8",
     "@ethersproject/bignumber": "^5.1.1",
     "@ethersproject/bytes": "^5.0.8",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.10",
     "ethers": "^5.1.0",
+    "next": "^13.2.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-feather": "^2.0.10",
-    "react-router-dom": "^6.4.3",
-    "react-scripts": "5.0.0",
     "react-tooltip": "^5.5.1",
     "typescript": "^4.4.2",
-    "wagmi": "^0.8.10",
-    "web-vitals": "^1.0.1",
-    "web3modal": "^1.9.4"
+    "wagmi": "^0.8.10"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "bundlesize": "react-scripts build --stats && webpack-bundle-analyzer ./build/bundle-stats.json",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "lint": "eslint 'src/**/*.{js,ts,tsx}'",
     "prettier:format": "yarn prettier './**/*.{js,json,md,ts,tsx,yml}' --write && yarn run lint --fix"
   },
