--- src/index.css
@@ -1,66 +1,0 @@
-.header-logo {
-  width: 100%;
-}
-
-.header-logo img {
-  height: 300px;
-  filter: drop-shadow(2px 4px 26px rgba(255,255,255,0.4));
-  pointer-events: none;
-}
-
-@media screen and (max-width: 400px) {
-  .header-logo img {
-    height: 240px;
-  }
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .header-logo img {
-    animation: Header-logo-spin infinite 20s linear;
-  }
-}
-
-header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  color: white;
-}
-
-header > div {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  margin-top: var(--space-l);
-}
-
-.receipt-text {
-  margin: var(--space-m) auto;
-}
-
-@keyframes Header-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.external-link {
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  font-size: var(--font-size-m);
-}
-
-.text-align-center {
-  text-align: center;
-}
-
-.react-tooltip {
-  font-size: var(--font-size-m);
-}
