+++ src/pages/index.tsx
@@ -1,0 +1,7 @@
+import { NextPage } from 'next';
+
+const Index: NextPage = () => {
+  return null;
+};
+
+export default Index;
